tests/wpt/tests/css/css-anchor-position/container-queries/container-type-anchored-computed.html

<!DOCTYPE html>
<title>CSS Conditional Test: Computed values of container-type with 'anchored'</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-2/#anchored-container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
  test_computed_value('container-type', 'anchored');
  test_computed_value('container-type', 'anchored size', 'size anchored');
  test_computed_value('container-type', 'inline-size anchored');
</script>
